requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
flask==2.2.5
